pyproject.toml
[project]
name = "tic-tac-toe-rl"
version = "0.1.0"
description = "Playing around with reinforcement learning in tic-tac-toe"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chex==0.1.87",
    "cupy-cuda12x>=13.3.0",
    "cython>=3.0.11",
    "flax>=0.7.4",
    "gymnasium>=0.29.1",
    "gymnax>=0.0.8",
    "jupyterlab>=4.2.5",
    "setuptools>=74.1.2",
    "tensordict>=0.6.1",
    "torch>=2.5.1",
    "torchrl>=0.6.0",
    "tqdm>=4.66.6",
    "tyro>=0.8.14",
]
